fix create user automation test
